+++ crates/core/src/lib.rs
@@ -163,7 +163,7 @@ impl FixedTimestepSystem {
     }
 }
 impl System for FixedTimestepSystem {
-    #[ambient_profiling::function]
+    #[profiling::function]
     fn run(&mut self, world: &mut World, event: &FrameEvent) {
         let delta_time = *world.resource(self::delta_time());
         self.acc += delta_time;