+++ crates/physics/src/lib.rs
@@ -233,7 +233,7 @@ pub fn server_systems() -> SystemGroup {
 /// Results will be available after [`fetch_simulation_system`]
 pub fn run_simulation_system() -> DynSystem {
     Box::new(FnSystem::new(|world, _| {
-        ambient_profiling::scope!("run_simulation_system");
+        profiling::scope!("run_simulation_system");
         let scene = world.resource(main_physics_scene());
         scene.simulate(FIXED_SERVER_TICK_TIME.as_secs_f32());
     }))
@@ -244,7 +244,7 @@ pub fn run_simulation_system() -> DynSystem {
 /// Must only be called once per [`run_simulation_system`]
 pub fn fetch_simulation_system() -> DynSystem {
     Box::new(FnSystem::new(|world, _| {
-        ambient_profiling::scope!("fetch_simulation_system");
+        profiling::scope!("fetch_simulation_system");
 
         world.resource(collisions()).lock().clear();
         world.resource_mut(collider_loads()).clear();